+++ src/main/java/com/luoying/core/old/JavaDockerCodeSandBox_Old.java
@@ -191,7 +191,7 @@ public class JavaDockerCodeSandBox_Old implements CodeSandBox {
             };
 
             // 获取占用的内存
-            final long[] maxMemory = {0L};
+            final double[] maxMemory = {0L};
             StatsCmd statsCmd = dockerClient.statsCmd(containerId);
             ResultCallback<Statistics> statisticsResultCallback = statsCmd.exec(new ResultCallback<Statistics>() {
                 @Override
