+++ src/main/java/com/luoying/model/QuestionSubmitJudgeInfo.java
@@ -2,6 +2,10 @@ package com.luoying.model;
 
 import lombok.Data;
 
+/**
+ * @author 落樱的悔恨
+ * 判题信息
+ */
 @Data
 public class QuestionSubmitJudgeInfo {
     /**
@@ -12,10 +16,10 @@ public class QuestionSubmitJudgeInfo {
     /**
      * 内存消耗 KB
      */
-    private Long memory;
+    private Double memory;
 
     /**
-     * 时间消耗 KB
+     * 时间消耗 ms
      */
     private Long time;
 }
